Erstatt liste med TilkjentPeriode med tidslinje
