+++ app/src/main/kotlin/no/nav/aap/api/postgres/BehandlingsRepository.kt
@@ -1,14 +1,8 @@
 package no.nav.aap.api.postgres
 
-import com.papsign.ktor.openapigen.annotations.properties.description.Description
 import java.math.BigDecimal
-import java.math.RoundingMode
-import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.LocalDateTime
-import kotlin.math.roundToInt
-import no.nav.aap.api.intern.Kilde
-import no.nav.aap.api.intern.VedtakUtenUtbetaling
 import no.nav.aap.api.kelvin.Avslagsårsak
 import no.nav.aap.api.kelvin.Behandling
 import no.nav.aap.api.kelvin.GjeldendeStansEllerOpphør
@@ -17,9 +11,11 @@ import no.nav.aap.api.kelvin.KelvinSakStatus
 import no.nav.aap.api.kelvin.RettighetsTypePeriode
 import no.nav.aap.api.kelvin.Sak
 import no.nav.aap.api.kelvin.StansEllerOpphør
-import no.nav.aap.api.kelvin.TilkjentPeriode
+import no.nav.aap.api.kelvin.TilkjentYtelse
 import no.nav.aap.api.kelvin.UnderveisIntern
 import no.nav.aap.komponenter.dbconnect.DBConnection
+import no.nav.aap.komponenter.tidslinje.Segment
+import no.nav.aap.komponenter.tidslinje.Tidslinje
 import no.nav.aap.komponenter.type.Periode
 import org.slf4j.LoggerFactory
 
@@ -212,19 +208,19 @@ class BehandlingsRepository(private val connection: DBConnection) {
                                               BARNETILLEGG, UFOREGRADERING)
                 VALUES (?, ?::daterange, ?, ?, ?, ?, ?, ?, ?, ?)
             """.trimIndent(),
-            behandling.tilkjent
+            behandling.tilkjent.segmenter()
         ) {
             setParams {
                 setLong(1, nytilkjentId)
-                setPeriode(2, Periode(it.tilkjentFom, it.tilkjentTom))
-                setBigDecimal(3, it.dagsats.toBigDecimal())
-                setInt(4, it.gradering)
-                setBigDecimal(5, it.grunnlagsfaktor)
-                setBigDecimal(6, it.grunnbeløp)
-                setInt(7, it.antallBarn)
-                setBigDecimal(8, it.barnetilleggsats)
-                setBigDecimal(9, it.barnetillegg)
-                setInt(10, it.samordningUføregradering)
+                setPeriode(2, it.periode)
+                setBigDecimal(3, it.verdi.dagsats.toBigDecimal())
+                setInt(4, it.verdi.gradering)
+                setBigDecimal(5, it.verdi.grunnlagsfaktor)
+                setBigDecimal(6, it.verdi.grunnbeløp)
+                setInt(7, it.verdi.antallBarn)
+                setBigDecimal(8, it.verdi.barnetilleggsats)
+                setBigDecimal(9, it.verdi.barnetillegg)
+                setInt(10, it.verdi.samordningUføregradering)
             }
         }
         if (behandling.beregningsgrunnlag != null) {
@@ -417,7 +413,7 @@ class BehandlingsRepository(private val connection: DBConnection) {
         }.toSet()
     }
 
-    private fun hentAvslagsårsaker(stansOpphørVurderingId: Long): Set<Avslagsårsak>{
+    private fun hentAvslagsårsaker(stansOpphørVurderingId: Long): Set<Avslagsårsak> {
         return connection.queryList(
             """SELECT * FROM avslagsarsak WHERE stans_opphor_vurdering_id = ?""".trimIndent()
         ){
@@ -430,7 +426,7 @@ class BehandlingsRepository(private val connection: DBConnection) {
         }.toSet()
     }
 
-    private fun hentTilkjentYtelse(behandlingId: Long): List<TilkjentPeriode> {
+    private fun hentTilkjentYtelse(behandlingId: Long): Tidslinje<TilkjentYtelse> {
         return connection.queryList(
             """
                 SELECT * FROM TILKJENT_PERIODE
@@ -442,20 +438,21 @@ class BehandlingsRepository(private val connection: DBConnection) {
         ) {
             setParams { setLong(1, behandlingId) }
             setRowMapper {
-                TilkjentPeriode(
-                    tilkjentFom = it.getPeriode("PERIODE").fom,
-                    tilkjentTom = it.getPeriode("PERIODE").tom,
-                    dagsats = it.getBigDecimal("DAGSATS").toInt(),
-                    gradering = it.getInt("GRADERING"),
-                    grunnlagsfaktor = it.getBigDecimal("GRUNNLAGSFAKTOR"),
-                    grunnbeløp = it.getBigDecimal("GRUNNBELOP"),
-                    antallBarn = it.getInt("ANTALL_BARN"),
-                    barnetilleggsats = it.getBigDecimal("BARNETILLEGGSATS"),
-                    barnetillegg = it.getBigDecimal("BARNETILLEGG"),
-                    samordningUføregradering = it.getIntOrNull("UFOREGRADERING")
+                Segment(
+                    it.getPeriode("PERIODE"),
+                    TilkjentYtelse(
+                        dagsats = it.getBigDecimal("DAGSATS").toInt(),
+                        gradering = it.getInt("GRADERING"),
+                        grunnlagsfaktor = it.getBigDecimal("GRUNNLAGSFAKTOR"),
+                        grunnbeløp = it.getBigDecimal("GRUNNBELOP"),
+                        antallBarn = it.getInt("ANTALL_BARN"),
+                        barnetilleggsats = it.getBigDecimal("BARNETILLEGGSATS"),
+                        barnetillegg = it.getBigDecimal("BARNETILLEGG"),
+                        samordningUføregradering = it.getIntOrNull("UFOREGRADERING")
+                    )
                 )
             }
-        }
+        }.let { Tidslinje(it) }
     }
 
     fun erNyttVedtak(fnr: String): Boolean {
@@ -517,87 +514,3 @@ data class BehandlingDB(
     val vedtakId: Long? = null,
     val førstegangsbehandling: Boolean
 )
-
-/**
- * @param uføregrad Svarer til prosent uføre.
- */
-data class TilkjentDB(
-    val dagsats: Int,
-    val gradering: Int,
-    val grunnlagsfaktor: BigDecimal,
-    val grunnbeløp: BigDecimal,
-    val antallBarn: Int,
-    val barnetilleggsats: BigDecimal,
-    val barnetillegg: BigDecimal,
-    val uføregrad: Int? = 0,
-) {
-    fun gradertBarnetillegg(): BigDecimal =
-        this.barnetillegg.multiply(
-            this.gradering.toBigDecimal()
-                .divide(100.toBigDecimal())
-        ).setScale(0, RoundingMode.HALF_UP)
-
-    fun regnUtDagsatsEtterUføreReduksjon(): Int =
-        this.dagsats.times(
-            (100 - (this.uføregrad ?: 0)) / 100.0
-        ).roundToInt()
-}
-
-fun weekdaysBetween(startDate: LocalDate, endDate: LocalDate): Int {
-    var count = 0
-    var date = startDate
-
-    while (!date.isAfter(endDate)) {
-        if (date.dayOfWeek != DayOfWeek.SATURDAY && date.dayOfWeek != DayOfWeek.SUNDAY) {
-            count++
-        }
-        date = date.plusDays(1)
-    }
-
-    return count
-}
-
-/**
- * @param vedtakId Svarer til ID til vedtak-tabellen i behandlingsflyt.
- */
-data class VedtakUtenUtbetalingUtenPeriode(
-    val vedtakId: String,
-    @param:Description("Full dagsats før reduksjoner.")
-    val dagsats: Int,
-    @param:Description("Dagsats etter uføre-reduksjon. Dette er lik dagsats * (100 - uføregrad) / 100. Kommer kun fra nytt system (Kelvin). Ved manglende data er denne null.")
-    val dagsatsEtterUføreReduksjon: Int,
-    @param:Description("Status på et vedtak. Mulige verdier er LØPENDE, AVSLUTTET, UTREDES.")
-    val status: String, //Hypotese, vedtaksstatuskode
-    val saksnummer: String,
-    val vedtaksdato: LocalDate, //reg_dato
-    @param:Description("Rettighetsgruppe. For data fra Arena er dette aktivitetsfasekode.")
-    val rettighetsType: String, ////aktivitetsfase //Aktfasekode
-    val beregningsgrunnlag: Int,
-    val barnMedStonad: Int,
-    @param:Description("Kildesystem for vedtak. Mulige verdier er ARENA og KELVIN.")
-    val kildesystem: Kilde,
-    val samordningsId: String? = null,
-    val opphorsAarsak: String? = null,
-    val barnetilleggSats: BigDecimal? = null,
-) {
-    fun tilVedtakUtenUtbetaling(periode: no.nav.aap.api.intern.Periode): VedtakUtenUtbetaling {
-        return VedtakUtenUtbetaling(
-            vedtakId = this.vedtakId,
-            dagsats = this.dagsats,
-            dagsatsEtterUføreReduksjon = this.dagsatsEtterUføreReduksjon,
-            status = this.status,
-            saksnummer = this.saksnummer,
-            vedtaksdato = this.vedtaksdato,
-            vedtaksTypeKode = null,
-            vedtaksTypeNavn = null,
-            periode = periode,
-            rettighetsType = this.rettighetsType,
-            beregningsgrunnlag = this.beregningsgrunnlag,
-            barnMedStonad = this.barnMedStonad,
-            barnetillegg = barnMedStonad * (this.barnetilleggSats?.toInt() ?: 0),
-            kildesystem = this.kildesystem,
-            samordningsId = this.samordningsId,
-            opphorsAarsak = this.opphorsAarsak
-        )
-    }
-}
