+++ app/src/test/kotlin/no/nav/aap/api/postgres/TilkjentYtelseTest.kt
@@ -1,12 +1,13 @@
 package no.nav.aap.api.postgres
 
+import no.nav.aap.api.kelvin.TilkjentYtelse
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
-class TilkjentDBTest {
+class TilkjentYtelseTest {
     @Test
     fun `gradering etter uføre`() {
-        val tilkjent = TilkjentDB(
+        val tilkjent = TilkjentYtelse(
             dagsats = 1000,
             gradering = 50,
             grunnlagsfaktor = 2.4.toBigDecimal(),
@@ -14,7 +15,7 @@ class TilkjentDBTest {
             antallBarn = 0,
             barnetilleggsats = 37.toBigDecimal(),
             barnetillegg = 0.toBigDecimal(),
-            uføregrad = 30
+            samordningUføregradering = 30
         )
 
         val dagsatsEtterUføreReduksjon = tilkjent.regnUtDagsatsEtterUføreReduksjon()
